add the regex and add some modifactions to some functions

+++ proj.sh
@@ -2,14 +2,19 @@
 let cont=0
 let match=0
 let connect=0
-
+let coundel=0
+let coundeltable=0
 # Define the db_menu function
 db_menu() {
-    select t in Create-Table List-Tables Drop-Table Insert-into-Table Select-From-Table Delete-From-Table Update-Table Back-to-Main-Menu; do
+    select t in Create-Table List-Tables Drop-Table Insert-into-Table Select-From-Table Delete-From-Table Update-Table Back-to-Main-Menu
+    do
         case $t in
             'Create-Table')
-                read -p "Enter the name of your table: " table
+                read -p "Enter the name of your table with this pattern(A-Za-z]+$): " table
                 ct=0
+                regxCtable='^[A-Za-z]+$'
+	if [[ $table =~ $regxCtable ]]
+	then
                 for i in $(ls); do
                     if [[ $i == $table ]]; then
                         ct+=1
@@ -17,9 +22,12 @@ db_menu() {
                 done
 
                 if [[ ct -gt 0 ]]; then
-                    echo "This table already exists."
+                    echo "---->This table already exists "
                 else
                     touch $table
+                fi
+                else
+                echo "---->invalid pattern of name's table "
                 fi
                 ;;
             'List-Tables')
@@ -27,16 +35,68 @@ db_menu() {
                 ;;
             'Drop-Table')
                 read -p "Enter the name of the table to delete: " table
-                rm $table
+                echo "$(ls $path)"
+                coundeltable=0
+                for i in $(ls $path)
+					do
+						if [[ $table == $i ]]
+						then
+						coundeltable+=1
+						fi
+					done
+				if [[ $coundeltable -gt 0 ]]
+				then
+				     rm $table
+				    echo "the table of $table has been deleted."
+				else
+				   echo " this $table is not found to delet"
+				fi
                 ;;
             'Insert-into-Table')
-                read -p "Enter the name of the table: " table
+                read -p "Enter the name of the table with this pattern (^[A-Za-z]+$) : " table
+                 regxNtable='^[A-Za-z]+$'
+	      if [[ $table =~ $regxNtable ]]
+	      then
                 if [[ -f $table ]]; then
-                    read -p "Insert name of columns of table: " columns
-                    read -p "Insert values: " values
-                    echo "$columns : $values" >> $table
+                    read -p "Insert name of columns of table with this pattern (^[A-Za-z]+$) : " columns
+                    regxcol='^[A-Za-z]+$'
+	          if [[ $columns =~ $regxcol ]]
+	          then
+	          
+                    read -p "Insert data-type of columns of table string or number: " datatypecol
+		         if [[ $datatypecol == "string" ]]
+		         then
+		          read -p "Insert string values with this pattern (^[A-Za-z]+$) : " values1
+		            regxval1='^[A-Za-z]+$'
+				if [[ $values1 =~ $regxval1 ]]
+				then
+			
+		                     echo "$columns : $values1" >> $table
+		                     else
+		                     echo "invalid pattern of string value"
+		                     fi
+		          elif [[ $datatypecol == "number" ]]
+		           then
+		            read -p "Insert numerical values with this pattern (^[0-9]+$)  : " values2
+		                regxval2='^[0-9]+$'
+				if [[ $values2 =~ $regxval2 ]]
+				then
+		                     echo "$columns : $values2" >> $table
+		                     else
+		                     echo "invalid pattern of numerical value"
+		                     fi
+		          else
+		          echo "--------error : please enter right datatype >string Or number only-----"
+		          fi
+		 
+		      else
+		      echo "invalid pattern of column"
+		      fi
                 else
                     echo "Table $table does not exist."
+                fi
+                else
+                echo "invalid pattern name of table"
                 fi
                 ;;
             'Select-From-Table')
@@ -67,6 +127,7 @@ db_menu() {
                 fi
                 ;;
             'Back-to-Main-Menu')
+            cd $HOME/DBSQL
                 break
                 ;;
             *)
@@ -95,6 +156,9 @@ if [[ $cont -gt 0 ]]
 				'create-DB')
 					read -p '-------------------------------create your DB : ' schema
 					match=0
+					regxC='^[a-zA-Z]+[0-9]+$'
+					       if [[ $schema =~ $regxC ]]
+						then
 					for i in `ls $path`
 					do
 					         if [[ $schema == $i ]] 
@@ -106,8 +170,12 @@ if [[ $cont -gt 0 ]]
 						         then
 							   mkdir $schema
 							   else
-							   echo "this schema is founded " 
-						fi  
+							   echo "this schema is already found " 
+						fi 
+						    else 
+	echo "invalid pattern of name"
+	fi
+						 
 					;;
 				'list-DB')
 				echo "-------------------------------list all DB"
@@ -115,7 +183,21 @@ if [[ $cont -gt 0 ]]
 					;;
 				'drop-DB')
 				read -p "-------------------------------delet any DB : " del
-					rm -r $del
+				 coundel=0
+					for i in $(ls $path)
+					do
+						if [[ $del == $i ]]
+						then
+						coundel=+1
+						fi
+					done
+					if [[ $coundel -gt 0 ]]
+					then
+					   rm -r $del
+					   echo "---->Database $del has been deleted."
+					else
+					     echo "---->this DB not found to delet"
+					fi
 					;;
 				'connect-DB')
 				read -p "-------------------------------connect your DB : " conn
@@ -125,7 +207,7 @@ if [[ $cont -gt 0 ]]
 						if [[  $i == $conn ]]
 						then
 						connect+=1
-						echo "$conn will connect now"
+						echo "$conn is connected now"
 						fi
 					done
 					if [[ $connect -gt 0 ]] 
@@ -145,10 +227,3 @@ if [[ $cont -gt 0 ]]
 else
 	mkdir DBSQL
 fi
-<<COMMENT
-for i in ls 
-do
-if [[ schema == i ]] 
-done
-path= cd ../DBSQL 
-COMMENT
